Restructure birthday, allow setting timezone if unknown
